pyproject.toml
[tool.poetry]
name = "codeprompt"
version = "0.1.0"
description = ""
authors = ["Nutchanon Ninyawee <nutchanon@codustry.com>"]
readme = "README.md"

[tool.poetry.scripts]
codeprompt = "codeprompt.codeprompt:main"

[tool.poetry.dependencies]
python = "^3.9"
pydantic = "^1.10.7"
click = "^8.1.3"
pygithub = "^1.58.1"
pyperclip = "^1.8.2"
jinja2 = "^3.1.2"
python-frontmatter = "^1.0.0"
tiktoken = "^0.3.3"
loguru = "^0.7.0"


[tool.poetry.group.dev.dependencies]
mypy = "^1.2.0"
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
